Restructure several folders
