+++ src/components/Header/components/Hamburger/index.js
@@ -1,0 +1,12 @@
+import React from "react";
+import styles from "./styles.scss";
+
+const Hamburger = ({ handleClick }) => (
+  <div className="toggle-container">
+    <div className="bar"></div>
+    <div className="bar"></div>
+    <div className="bar"></div>
+  </div>
+);
+
+export default Hamburger;
