+++ src/components/Header/components/Menu/Menu.module.scss
@@ -1,5 +1,5 @@
-@import "../../styles/mixins.scss";
-@import "../../styles/variables.scss";
+@import "../../../../styles/mixins.scss";
+@import "../../../../styles/variables.scss";
 
 .wrapper {
   @include flexContainer;
