+++ src/components/Header/components/Navigation/Navigation.module.scss
@@ -1,4 +1,4 @@
-@import "../../styles/mixins.scss";
+@import "../../../../styles/mixins.scss";
 
 .navigation {
   grid-area: navigation;
