+++ src/styles/mixins.scss
@@ -4,22 +4,48 @@
   max-height: 100%;
   max-width: 100%;
 }
+
 @mixin centerAlignment {
   align-items: center;
   justify-content: center;
 }
+
 @mixin flexContainer {
   display: flex;
   position: relative;
   @include centerAlignment;
 }
+
 @mixin gridContainer {
   display: grid;
   @include centerAlignment;
 }
+
 @mixin text {
   letter-spacing: 0.5px;
   font-size: 14px;
   font-weight: 700;
   color: var(--gray-800);
+}
+
+$breakpoints: (
+  "xs": 320,
+  "sm": 600,
+  "md": 1024,
+  "lg": 1280,
+  "xl": 1440
+) !default;
+
+@mixin responsive($breakpoint: "lg", $type: "max") {
+  @if map-has-key($breakpoints, $breakpoint) {
+    @if $type == "min" {
+      @media (min-width: map-get($breakpoints, $breakpoint) + "px") {
+        @content;
+      }
+    } @else {
+      @media (max-width: map-get($breakpoints, $breakpoint) - 1 + "px") {
+        @content;
+      }
+    }
+  }
 }
